+++ backend/routes/lesson-edit.js
@@ -1,10 +1,10 @@
 const getContent = require("../githubAPI/getRepoContent");
-const saveFile = require("../utils/save-file");
-const upsertUserLessons = require("../utils/upsert-user-lessons");
+const saveFile = require("../storage/save-file");
+const upsertUserLessons = require("../lesson/upsert-user-lessons");
 const axios = require("axios");
-const downloadImage = require("../utils/download-image");
-const isLessonSaved = require("../utils/is-lesson-saved");
-const createNewLesson = require("../utils/create-lesson-data");
+const downloadImage = require("../storage/download-image");
+const isLessonSaved = require("../lesson/is-lesson-saved");
+const createNewLesson = require("../lesson/create-lesson-data");
 const paths = require("../paths");
 const imgRegex = /([a-zA-Z0-9\s_\\.\-:])+(.png|.jpeg|.jpg)$/i;
 const mdRegex = /([a-zA-Z0-9\s_\\.\-:])+(.md)$/i;
