+++ backend/routes/lesson-files.js
@@ -1,5 +1,5 @@
 const paths = require("../paths");
-const listFiles = require("../utils/list-files");
+const listFiles = require("../lesson/list-files");
 module.exports = (app) => {
   app.get(paths.LESSON_FILES, async (req, res) => {
     const { lessonId } = req.params;
