+++ backend/storage/save-to-disk.spec.js
@@ -1,6 +1,6 @@
 const fs = require("fs");
 const saveToDisk = require("./save-to-disk");
-const getTempDir = require("./get-temp-dir");
+const getTempDir = require("../utils/get-temp-dir");
 
 it("should save file to disk", async () => {
   const filename = getTempDir(["just", "for", "test", "testfile.txt"]);
