+++ backend/thumb/thumb-fetch.js
@@ -1,7 +1,7 @@
 const thumbUrl = require("./thumb-url");
 const saveFile = require("../storage/save-file");
-const isAppEngine = require("./isAppEngine");
-const gcsUrl = require("./gcs-url");
+const isAppEngine = require("../utils/isAppEngine");
+const gcsUrl = require("../utils/gcs-url");
 const downloadImage = require("../storage/download-image");
 module.exports = async (previewUrl, storagePath) => {
   const url = thumbUrl(previewUrl);
