+++ examples/list-files.js
@@ -1,4 +1,4 @@
-const listFiles = require("../backend/utils/list-files");
+const listFiles = require("../backend/lesson/list-files");
 
 const examples = async () => {
   const files = await listFiles(["mijohansen", "microbit", "pxt_gangespill"]);
