+++ examples/load-from-gcs.js
@@ -1,4 +1,4 @@
-const loadToString = require("../backend/utils/load-file-gcs-to-string");
+const loadToString = require("../backend/storage/load-file-gcs-to-string");
 
 loadToString("users/mijohansen/lessons.json", "lessoneditor").then((content) =>
   console.log(content)
